refactor(scene-engine): convert scene-engine/shared imports to relative paths
