+++ frontend/src/features/scene-engine/steps/character/extraction/components/ReferenceImageViewer.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { Skeleton } from "../../../../components/Skeleton";
 import { imageSignedUrlOptions } from "../../character.queries";
-import type { ImageRecord } from "@/features/story/shared/story.types";
+import type { ImageRecord } from "../../../../shared/scene-engine.types";
 
 type ReferenceImageViewerProps = {
   img: ImageRecord;
