+++ frontend/src/features/scene-engine/steps/character/rendering/components/CharacterRenderingList.tsx
@@ -3,7 +3,7 @@ import { Skeleton } from "../../../../components/Skeleton";
 import { useSceneEngine } from "../../../../context";
 import { imageSignedUrlOptions } from "../../character.queries";
 import type { CharacterBundle } from "../../character.types";
-import type { ImageRecord } from "@/features/story/shared/story.types";
+import type { ImageRecord } from "../../../../shared/scene-engine.types";
 import { useCharacterRendering } from "../hooks/useCharacterRendering";
 
 type CharacterRenderBlockProps = {
