+++ frontend/src/features/story/shared/story.constants.ts
@@ -1,0 +1,2 @@
+export const STORY_API_BASE = "/api/comic-builder/v2" as const;
+export const STORY_QUERY_ROOT = ["story"] as const;
